% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{bootVar.multiallelic}
\alias{bootVar.multiallelic}
\title{get variance from the bootstrap for multi-allelic site;}
\usage{
bootVar.multiallelic(ustat.meta, vstat.meta, r2, nStudy, nBoot = 100)
}
\arguments{
\item{ustat.meta}{score statistic from meta-analysis}

\item{vstat.meta}{standard deviation of the meta-analysis score statistics}

\item{r2}{the LD matrix;}

\item{nStudy}{the number of studies;}
}
\description{
get variance from the bootstrap for multi-allelic site;
}
