% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.single.R
\name{bootVar}
\alias{bootVar}
\title{bootVar}
\usage{
bootVar(dat, r2, nBoot = 1000, ix.candidate, ix.known, N, ix.rp)
}
\arguments{
\item{dat}{the summary statistic calculated from the data;}

\item{nBoot}{default is 100;}

\item{ix.candidate}{the index of the candidate variants;}

\item{ix.known}{the index of the known variants;}
}
\value{
estimated conditional variance
}
\description{
bootVar
}
