% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.TWAS.R
\name{calc.r2}
\alias{calc.r2}
\title{calculate R2 from a list of variants;}
\usage{
calc.r2(vcf.ref.file, tabix.range, refGeno = "GT")
}
\arguments{
\item{vcf.ref.file}{VCF reference file}

\item{tabix.range}{tabix range for a list of variants;}

\item{refGeno}{the FORMAT field for genotypes; could be GT or DS;}
}
\value{
calcualted r2 matrix; with positions being the row and colnames;
}
\description{
calculate R2 from a list of variants;
}
