% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.utils.R
\name{cauchy.p}
\alias{cauchy.p}
\title{combine p-values from Cauchy distribution;}
\usage{
cauchy.p(pval, weight = NULL)
}
\arguments{
\item{pval}{a vector of p-values}

\item{weight}{a vector of weights, the default is equal weight, and the weights add up to 1}
}
\value{
combined p-values;
}
\description{
combine p-values from Cauchy distribution;
}
