% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{convertChisq2Beta}
\alias{convertChisq2Beta}
\title{convert chisq statistic to beta for binary trait assuming the variance for y;}
\usage{
convertChisq2Beta(statistic, beta.ori, var.y, af, N, binaryTrait = FALSE)
}
\arguments{
\item{statistic}{chisq stat;}

\item{beta.ori}{original beta; for getting the sign;}

\item{var.y}{the variance for Y. for binary trait, the variance of y is frac.case*(1-frac.case); for continuous trait, the variance for y is typically set to 1;}

\item{af}{allele freq;}

\item{N}{sample size;}
}
\description{
convert chisq statistic to beta for binary trait assuming the variance for y;
}
