% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.utils.R
\name{cor2cov}
\alias{cor2cov}
\title{a faster version to convert correlation matrix to covariance matrix;}
\usage{
cor2cov(cor.mat, sd.vec)
}
\arguments{
\item{cor.mat}{correlation matrix;}

\item{sd.vec}{vector of standard deviation;}
}
\description{
a faster version to convert correlation matrix to covariance matrix;
}
