% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{emp.copula}
\alias{emp.copula}
\title{estimate the empirical correlation between statistics;}
\usage{
emp.copula(statistic.mat, p.value.mat)
}
\arguments{
\item{statistic.mat}{matrix of statistics}

\item{p.value.mat}{matrix of p-values that corresponds to the statistic}
}
\value{
p.value.minp and the estimated covariance matrix;
}
\description{
estimate the empirical correlation between statistics;
}
