% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.gene.core.R
\name{estimateH2}
\alias{estimateH2}
\title{Estimate the variance explained by the variants in a locus;}
\usage{
estimateH2(...)
}
\arguments{
\item{r2.rare}{the r2 matrix for rare variants;}

\item{maf.rare}{the maf vector for rare variants;}
}
\value{
r2 estimates;
}
\description{
Estimate the variance explained by the variants in a locus;
}
