% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{estimateMAeffect.meta}
\alias{estimateMAeffect.meta}
\title{estimate multi-allelic effect from summary stat;}
\usage{
estimateMAeffect.meta(ustat.mat, vstat.mat, boot = FALSE, nBoot = 100)
}
\arguments{
\item{ustat.mat}{ustat;}

\item{vstat.mat}{ustat;}
}
\description{
estimate multi-allelic effect from summary stat;
}
