% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{estimateMAeffect}
\alias{estimateMAeffect}
\title{estimate multi-allelic effect from summary stat;}
\usage{
estimateMAeffect(ustat, vstat, af, pos, boot = FALSE, nBoot = 100)
}
\arguments{
\item{ustat}{ustat;}

\item{vstat}{ustat;}

\item{af}{af;}

\item{pos}{pos;}
}
\description{
estimate multi-allelic effect from summary stat;
}
