% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.twas.core.R
\name{estimatePredictedBeta.cor.fast}
\alias{estimatePredictedBeta.cor.fast}
\title{a faster way (with some approximations) to estimate the correlations between multiple TWAS statistics}
\usage{
estimatePredictedBeta.cor.fast(dat, varList, af.pca, af.pca.eqtl)
}
\arguments{
\item{dat}{dataset that contains of eQTL weight, score statistics, and covariance matrix}

\item{varList}{variant list as well as corresponding transcript;}

\item{af.pca}{allele frequency PCs}

\item{af.pca.eqtl}{PCs for the eqtl dataset}
}
\description{
a faster way (with some approximations) to estimate the correlations between multiple TWAS statistics
}
