% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.twas.core.R
\name{estimatePredictedBeta.cor}
\alias{estimatePredictedBeta.cor}
\title{estimate correlations between multiple twas statistics;}
\usage{
estimatePredictedBeta.cor(dat, varList, af.pca, af.pca.eqtl)
}
\arguments{
\item{dat}{dataset that contains of eQTL weight, score statistics, and covariance matrix}

\item{varList}{variant list as well as corresponding transcript;}

\item{af.pca}{allele frequency PCs}

\item{af.pca.eqtl}{PCs for the eqtl dataset}
}
\description{
estimate correlations between multiple twas statistics;
}
