% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.twas.core.R
\name{estimateTESLA.cor}
\alias{estimateTESLA.cor}
\title{estimate the correlation between minimal p-value statistic}
\usage{
estimateTESLA.cor(gcov.twas, no.pc = 3)
}
\arguments{
\item{no.pc}{no.pc;default is 3;}
}
\description{
estimate the correlation between minimal p-value statistic
}
