% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.twas.core.R
\name{estimateTWAS.cor}
\alias{estimateTWAS.cor}
\title{estimate TWAS Statistic Correlation}
\usage{
estimateTWAS.cor(dat, varList, af.pca, af.pca.eqtl, maxNumVar = 500)
}
\arguments{
\item{dat}{dataset;}

\item{varList}{variant list with genes being the first column and the variant positions (in the format of 1:123) in the second column;}

\item{af.pca}{MDS component for the GWAS dataset;}

\item{af.pca.eqtl}{MDS component for the eqtl dataset}
}
\description{
estimate TWAS Statistic Correlation
}
