% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.utils.R
\name{fread.big}
\alias{fread.big}
\title{check if the variable exist before reading large files using fread;}
\usage{
fread.big(...)
}
\arguments{
\item{varName}{variable name;}

\item{header}{if the file contain header default is NO;}

\item{file}{the file name to be read;}
}
\description{
check if the variable exist before reading large files using fread;
}
