% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.utils.R
\name{get.precise.pval}
\alias{get.precise.pval}
\title{get precise value for highly significant p-values;}
\usage{
get.precise.pval(statistic, df = 1)
}
\arguments{
\item{statistic}{the chisq statistic;}
}
\value{
a vector of character-based p-values;
}
\description{
get precise value for highly significant p-values;
}
