% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.single.R
\name{getCondUV.sand}
\alias{getCondUV.sand}
\title{get conditional U and V for sandwich estimator;}
\usage{
getCondUV.sand(...)
}
\arguments{
\item{ix.candidate}{candidate variant indices;}

\item{ix.known}{conditioned variant indices}

\item{dat}{the list with contributed summary stat;}

\item{r2}{LD matrices;}

\item{n.r2}{sample size for the ref panel;}
}
\value{
a list with condiitonal association analysis statistics;
}
\description{
get conditional U and V for sandwich estimator;
}
