% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.single.R
\name{getCondUV.wrapper}
\alias{getCondUV.wrapper}
\title{wrapper of getCondUV.sand for trans-ethnic and pooled analysis;}
\usage{
getCondUV.wrapper(
  dat,
  lambda,
  ix.candidate,
  ix.known,
  r2,
  sandwich,
  boot = TRUE,
  nBoot = 100,
  r2.list,
  study.ancestry,
  trans.ethnic,
  af.pca = NULL,
  cor.z2 = NULL,
  otho = TRUE
)
}
\description{
wrapper of getCondUV.sand for trans-ethnic and pooled analysis;
}
