% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{getImpQual}
\alias{getImpQual}
\title{get imputation quality from tables based upon imputation quality files;}
\usage{
getImpQual(imp.qual, pos, col.impqual)
}
\arguments{
\item{imp.qual}{the list with imputation qualities}

\item{pos}{A vector of positions where the imputatoin quality will be retrieved;}

\item{col.impqual}{The column where the imputation quality will be retrieved;}
}
\value{
a matrix with imputaton qualities
}
\description{
get imputation quality from tables based upon imputation quality files;
}
