% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gramSchmitt.R
\name{gram.schmitt}
\alias{gram.schmitt}
\title{Gram-Schmitt othogonalization}
\usage{
gram.schmitt(A, ip, ...)
}
\arguments{
\item{A}{a matrix;}

\item{ip}{a bivariate function that serves as inner product;}
}
\value{

}
\description{
Gram-Schmitt othogonalization
}
