% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{GWAMA.formatData}
\alias{GWAMA.formatData}
\title{format data into matrices;}
\usage{
GWAMA.formatData(raw.data.all, raw.imp.qual, impQualWeight, impQual.lb,
  col.impqual)
}
\arguments{
\item{raw.data.all}{The read in from summary assoicaiton statistic files;}

\item{raw.imp.qual}{The raw imputation quality files;}

\item{impQualWeight}{If using imputation quality as weight}
}
\value{
a list of converted data for downstream meta-analysis;
}
\description{
format data into matrices;
}

