% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{GWAMA.rmMulti}
\alias{GWAMA.rmMulti}
\title{remove multi-allelic sites;}
\usage{
GWAMA.rmMulti(dat)
}
\arguments{
\item{dat}{The dataset;}
}
\value{
the data set without multi-allelic sites;
}
\description{
remove multi-allelic sites;
}
