% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAMA.single.R
\name{GWAMA.single}
\alias{GWAMA.single}
\title{Single variant meta-analysis}
\usage{
GWAMA.single(score.stat.file, range, alternative = c("two.sided", "greater",
  "less"))
}
\arguments{
\item{score.stat.file}{files of score statistics}

\item{range}{tabix range of variants to be analyzed}

\item{alternative}{alternative hypothesis to be specified}
}
\value{
a list consisting of results
}
\description{
Single variant meta-analysis
}
