% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{l.mr.0}
\alias{l.mr.0}
\title{restricted likelihood with no variance heterogeneity;}
\usage{
l.mr.0(pars.0, b.vec, se.vec, af.pca)
}
\arguments{
\item{pars.0}{the parameters for the gamma}

\item{b.vec}{genetic effect estimates}

\item{se.vec}{the standard deviation of the genetic effect estimates}

\item{af.pca}{the covariates to be included}
}
\value{
the likelihood values;
}
\description{
restricted likelihood with no variance heterogeneity;
}
