% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{l.mr}
\alias{l.mr}
\title{meta-regression likelihood with variance heterogeneity;}
\usage{
l.mr(pars, b.vec, se.vec, af.pca)
}
\arguments{
\item{pars}{the parameters for the gamma and the log variance}

\item{b.vec}{genetic effect estimates}

\item{se.vec}{the standard deviation of the genetic effect estimates}

\item{af.pca}{the covariates to be included}
}
\value{
the likelihood values;
}
\description{
meta-regression likelihood with variance heterogeneity;
}
