% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{multiAlleleAssoc}
\alias{multiAlleleAssoc}
\title{Perform multi-allelic association tests;}
\usage{
multiAlleleAssoc(pos, dat, corMultiAllele.mat)
}
\arguments{
\item{pos}{The position of multi-allelic site;}

\item{dat}{The formatted data;}

\item{corMultiAllele.mat}{the correlation matrix generated from multiple alleles;}
}
\value{
a matrix with formatted result;
}
\description{
Perform multi-allelic association tests;
}

