% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvt.R
\name{pmaxnormsq}
\alias{pmaxnormsq}
\title{p-value for minimal stat for MVN;}
\usage{
pmaxnormsq(x, mu, sigma)
}
\arguments{
\item{x}{is the square of the max statistic for MVN}

\item{mu}{mean value}

\item{sigma}{the covariance matrix;}
}
\description{
p-value for minimal stat for MVN;
}
