% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{pmixchisq}
\alias{pmixchisq}
\title{calculate the mixture chisq p-value;}
\usage{
pmixchisq(statistic, lambda, h = NULL)
}
\arguments{
\item{statistic}{the test statistic;}

\item{lambda}{mixture proportion}

\item{h}{multiplicity, and the default is rep(1,length(lambda))}
}
\value{
p values for mixture chisq distributed random variables;
}
\description{
calculate the mixture chisq p-value;
}
