% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvt.R
\name{pvt}
\alias{pvt}
\title{function to calculate VT p-values;}
\usage{
pvt(statistic, mu, sigma, alternative = c("greater", "less", "two.sided"))
}
\arguments{
\item{statistic}{the z-score statistic that corresponds to the minimal p-values;}

\item{mu}{the mean value of the MVT distribution;}

\item{sigm}{the covariance matrix}
}
\description{
function to calculate VT p-values;
}
