% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.utils.R
\name{rareGWAMA.abf}
\alias{rareGWAMA.abf}
\title{compute ABF from summary stat;}
\usage{
rareGWAMA.abf(beta, se, tau = 0.04)
}
\arguments{
\item{beta}{genetic effect estimates;}

\item{se}{the SE of the genetic effect estimates;}
}
\value{

}
\description{
compute ABF from summary stat;
}
