% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.conditional.twas.R
\name{rareGWAMA.conditional.gene}
\alias{rareGWAMA.conditional.gene}
\title{conduct conditional analysis for multiple genes in}
\usage{
rareGWAMA.conditional.gene(
  score.stat.file,
  imp.qual.file = NULL,
  vcf.ref.file,
  anno,
  gc = FALSE,
  ...
)
}
\arguments{
\item{score.stat.file}{the file names of score statistic files;}

\item{imp.qual.file}{the file names of imputation quality;}

\item{vcf.ref.file}{the file names of the reference panel file;}

\item{tabix.range}{Tabix range for variants;}
}
\value{
formatted data and assocation testing results;
}
\description{
conduct conditional analysis for multiple genes in
}
