% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.estH2.R
\name{rareGWAMA.estH2}
\alias{rareGWAMA.estH2}
\title{Estimate the variance explained;}
\usage{
rareGWAMA.estH2(pos, beta.vec, beta.se, vcf.ref.file, ...)
}
\arguments{
\item{pos}{a vector of positions;}

\item{beta.vec}{A vector of estimated genetic effects;}

\item{beta.se}{A vector of estimated genetic effects se;}

\item{vcf.ref.file}{the file name of the VCF file}
}
\value{
A list of analysis results;
}
\description{
Estimate the variance explained;
}
