% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.QC.R
\name{rareGWAMA.flip}
\alias{rareGWAMA.flip}
\title{flip allele and fix the files}
\usage{
rareGWAMA.flip(fnameIn, fnameOut, refaltList)
}
\arguments{
\item{fnameIn}{input file name;}

\item{fnameOut}{output file name;}

\item{refaltList}{the list of group files;}
}
\description{
flip allele and fix the files
}
