% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.gene.core.R
\name{rareGWAMA.formatGene}
\alias{rareGWAMA.formatGene}
\title{organize the formatted stat into analyzable format;}
\usage{
rareGWAMA.formatGene(dat, ...)
}
\arguments{
\item{dat}{formatted data;}

\item{maf.cutoff}{The minor allele frequency cutoff}
}
\value{
dat which consist of formatted data for gene-level tests
}
\description{
organize the formatted stat into analyzable format;
}
