% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.gene.core.R
\name{rareGWAMA.gene}
\alias{rareGWAMA.gene}
\title{Conduct approximate gene-level tests and estimate variance explained;}
\usage{
rareGWAMA.gene(score.stat.file, imp.qual.file = NULL, vcf.ref.file,
  tabix.range, ...)
}
\arguments{
\item{score.stat.file}{the file names of score statistic files;}

\item{imp.qual.file}{the file names of imputation quality;}

\item{vcf.ref.file}{the file names of the reference panel file;}

\item{tabix.range}{Tabix range for variants;}
}
\value{
formatted data and assocation testing results;
}
\description{
Conduct approximate gene-level tests and estimate variance explained;
}
