% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.utils.R
\name{rareGWAMA.hmp}
\alias{rareGWAMA.hmp}
\title{harmonic mean p combination}
\usage{
rareGWAMA.hmp(pval.mat)
}
\arguments{
\item{pval.mat}{matrix of p-values; each row being one set of p-values;}
}
\value{
a vector of combined p-values using hmp;
}
\description{
harmonic mean p combination
}
