% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{rareGWAMA.single.multiAllele}
\alias{rareGWAMA.single.multiAllele}
\title{gateway function for multi-allelic analysis in rareGWAMA}
\usage{
rareGWAMA.single.multiAllele(score.stat.file, imp.qual.file = NULL,
  tabix.range, alternative = "two.sided", col.impqual = 5,
  impQual.lb = 0.7, impQualWeight = FALSE)
}
\arguments{
\item{imp.qual.file}{The imputation quality file names;}

\item{tabix.range}{The tabix range for the variants to be analyzed;}

\item{alternative}{The alternative hypothesis; only the default choice two.sided is implemented now;}

\item{col.impqual}{The column number for the imputation quality files; The default choice is 5.}

\item{impQual.lb}{The lower bound for the imputation quality cutoff; Variants with imputation quality less than the cutoffs are labelled as missing and removed from the meta-analysis}

\item{impQualWeight}{Whether to apply the imputation quality based optimal weighting;}

\item{score}{stat.file The score statistics file names;}

\item{cor.multiallele.file}{The correlation information for multi-allelic sites;}
}
\value{
A list of formatted output;
}
\description{
gateway function for multi-allelic analysis in rareGWAMA
}
