% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.twas.R
\name{rareGWAMA.twas.predEff}
\alias{rareGWAMA.twas.predEff}
\title{trans-ethnic TWAS method;}
\usage{
rareGWAMA.twas.predEff(dat, af.pca, af.pca.eqtl)
}
\arguments{
\item{dat}{the data that contains everything needed;}
}
\value{
a bunch of results, such as p-values;
}
\description{
trans-ethnic TWAS method;
}
