% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.TWAS.R
\name{rareGWAMA.TWAS}
\alias{rareGWAMA.TWAS}
\title{perform single tissue twas and multiple tissue twas;}
\usage{
rareGWAMA.TWAS(score.stat.file, imp.qual.file, vcf.ref.file, transcript, ...)
}
\arguments{
\item{score.stat.file}{the file names of score statistic files;}

\item{imp.qual.file}{the file names of imputation quality;}

\item{vcf.ref.file}{the file names of the reference panel file;}

\item{transcript}{The list of transcript for twas;}
}
\value{
formatted data and assocation testing results;
}
\description{
perform single tissue twas and multiple tissue twas;
}
