% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{trans.ethnic.meta.core}
\alias{trans.ethnic.meta.core}
\title{Trans-ethnic meta-analysis with simple interface;}
\usage{
trans.ethnic.meta.core(
  beta.mat,
  se.mat,
  af.pca,
  cor.z2,
  otho = TRUE,
  re = FALSE,
  recalibrate = FALSE
)
}
\arguments{
\item{beta.mat}{genetic effect estimates (one row per variant, and one column per study);}

\item{se.mat}{matrix of standard errors for the beta.mat}

\item{af.pca}{PCA based upon allele frequencies}

\item{cor.z2}{correlation of z2;}

\item{otho}{if to othorgonalize the af pca in case studies contain missing data;}

\item{re}{if to include random effect to capture nonancestral heterogeniety}
}
\description{
Trans-ethnic meta-analysis with simple interface;
}
