% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{trans.ethnic.meta}
\alias{trans.ethnic.meta}
\title{trans-ethnic meta-analysis}
\usage{
trans.ethnic.meta(
  dat,
  af.pca,
  cor.z2,
  otho = TRUE,
  re = FALSE,
  recalibrate = FALSE
)
}
\arguments{
\item{dat}{the formatted data;}

\item{af.pca}{the PCs or MDS coordinate for the allele frequencies}
}
\value{
trans-ethnic results;
}
\description{
trans-ethnic meta-analysis
}
