% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.single.R
\name{uniq.allele}
\alias{uniq.allele}
\title{unique alternative alleles;}
\usage{
uniq.allele(x)
}
\arguments{
\item{x}{A vector of allele labels;}
}
\value{
A vector of unique alleles;
}
\description{
unique alternative alleles;
}
