% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.finemap.R
\name{zscore.abf}
\alias{zscore.abf}
\title{compute ABF from summary stat;}
\usage{
zscore.abf(z, tau = 10000)
}
\arguments{
\item{beta}{genetic effect estimates;}

\item{se}{the SE of the genetic effect estimates;}
}
\value{
@export
}
\description{
compute ABF from summary stat;
}
