% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.finemap.R
\name{zscore.finemap}
\alias{zscore.finemap}
\title{fine mapping main function;}
\usage{
zscore.finemap(
  pval,
  cov.z,
  gene.vec,
  max.iter = 3,
  cs.level = 0.9,
  alpha = 5e-08,
  tau = 10000
)
}
\arguments{
\item{pval}{a vector p-values;}

\item{cov.z}{the covariance matrix between converted z-scores;}

\item{gene.vec}{a vector of gene names;}

\item{max.iter}{the max number of iteration conducted; default is 3, which means a maximum of 3 causal genes;}

\item{cs.level}{credible set level; default is set to 90%;}

\item{alpha}{significant threshold used to determine indpendently associated genes; default is 5e-8;}

\item{tau}{prior effect sizes; default is 10000, equivalent to non-informative prior;}
}
\description{
fine mapping main function;
}
