% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGWAMA.finemap.R
\name{zscore2pip}
\alias{zscore2pip}
\title{convert z-score to posterior inclusion probability;}
\usage{
zscore2pip(z, cov.z, gene.vec, cs.level = 0.9, tau = 10000)
}
\arguments{
\item{z}{a vector of z-scores for the locus;}

\item{cov.z}{the covariance between z-scores; order has to be the same as in z;}

\item{gene.vec}{a vector of gene names that matches z;}

\item{cs.level}{credible set level default is 90%;}

\item{tau}{prior on the effect sizes. default is 10000 which is equivalent to a non-informative prior;}
}
\description{
convert z-score to posterior inclusion probability;
}
